Jupyter notebooks updated to work with latest code:
The examples folder is now renamed to jupyter_notebooks.

Fixed 4 bugs relating to str to unicode conversions in Highlight_Position_Range in
Logo.py, get_rgb in colors.py, and in validate_matrix in validate.py and in
get_color_dict also in colors.py. This was needed to ensure that logomaker works both
with python 2 and 3.

Tested all jupyter notebooks with python and 2 and python 3.